refactor: remove client manager
