--- test/fixtures/apps/hpmweb/app/controllers/home.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports = function* () {
-  const rows = yield this.app.mysql.query('show tables');
-  this.body = {
-    hasRows: rows.length > 0,
-  };
-};
