--- test/fixtures/apps/hpmweb/app/router.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports = function(app) {
-  app.get('/', app.controller.home);
-};
