--- test/fixtures/apps/hpmweb/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "npmweb"
-}
