--- test/fixtures/apps/mysqlapp-multi-client-new/config/config.js
@@ -1,27 +1,0 @@
-'use strict';
-
-exports.mysql = {
-  clients: {
-    db1: {
-      host: '127.0.0.1',
-      port: 3307,
-      user: 'root',
-      password: '',
-      database: 'test',
-    },
-    db2: {
-      host: '127.0.0.1',
-      port: 3307,
-      user: 'root',
-      password: '',
-      database: 'test',
-    },
-    db3: {
-      host: '127.0.0.1',
-      port: 3307,
-      user: 'root',
-      password: '',
-      database: 'test',
-    },
-  }
-};
