--- test/fixtures/apps/mysqlapp-multi-client-new/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "mysqlapp-multi-client-new"
-}
