+++ test/fixtures/apps/mysqlapp-multi-client-wrong/package.json
@@ -1,3 +1,3 @@
 {
-  "name": "mysqlapp"
+  "name": "mysqlapp-multi-client-wrong"
 }
