--- test/fixtures/apps/mysqlapp-multi-client/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "mysqlapp"
-}
