--- test/fixtures/apps/mysqlapp-wrong-pwd/app/controller/home.js
@@ -1,10 +1,0 @@
-'use strict';
-
-module.exports = function* () {
-  const users = yield* this.service.user.list(this);
-
-  this.body = {
-    status: 'success',
-    users: users,
-  };
-};
