--- test/fixtures/apps/mysqlapp-wrong-pwd/app/service/user.js
@@ -1,5 +1,0 @@
-'use strict';
-
-exports.list = function* (ctx) {
-  return yield ctx.app.mysql.query('select * from npm_auth');
-};
