--- test/fixtures/apps/mysqlapp-wrong-pwd/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "mysqlapp"
-}
