+++ test/prepare.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const execSync = require('child_process').execSync;
+
+execSync('mysql -uroot test < test/npm_auth.sql');
+console.log('create table success');
